+++ tools/cabana/chart/signalselector.cc
@@ -1,6 +1,16 @@
-// SeriesSelector
+#include "tools/cabana/chart/signalselector.h"
 
-SeriesSelector::SeriesSelector(QString title, QWidget *parent) : QDialog(parent) {
+#include <QCompleter>
+#include <QDialogButtonBox>
+#include <QGridLayout>
+#include <QLabel>
+#include <QLineEdit>
+#include <QPushButton>
+#include <QVBoxLayout>
+
+#include "tools/cabana/streams/abstractstream.h"
+
+SignalSelector::SignalSelector(QString title, QWidget *parent) : QDialog(parent) {
   setWindowTitle(title);
   QGridLayout *main_layout = new QGridLayout(this);
 
@@ -42,24 +52,24 @@ SeriesSelector::SeriesSelector(QString title, QWidget *parent) : QDialog(parent)
   msgs_combo->model()->sort(0);
   msgs_combo->setCurrentIndex(-1);
 
-  QObject::connect(msgs_combo, qOverload<int>(&QComboBox::currentIndexChanged), this, &SeriesSelector::updateAvailableList);
+  QObject::connect(msgs_combo, qOverload<int>(&QComboBox::currentIndexChanged), this, &SignalSelector::updateAvailableList);
   QObject::connect(available_list, &QListWidget::currentRowChanged, [=](int row) { add_btn->setEnabled(row != -1); });
   QObject::connect(selected_list, &QListWidget::currentRowChanged, [=](int row) { remove_btn->setEnabled(row != -1); });
-  QObject::connect(available_list, &QListWidget::itemDoubleClicked, this, &SeriesSelector::add);
-  QObject::connect(selected_list, &QListWidget::itemDoubleClicked, this, &SeriesSelector::remove);
+  QObject::connect(available_list, &QListWidget::itemDoubleClicked, this, &SignalSelector::add);
+  QObject::connect(selected_list, &QListWidget::itemDoubleClicked, this, &SignalSelector::remove);
   QObject::connect(add_btn, &QPushButton::clicked, [this]() { if (auto item = available_list->currentItem()) add(item); });
   QObject::connect(remove_btn, &QPushButton::clicked, [this]() { if (auto item = selected_list->currentItem()) remove(item); });
   QObject::connect(buttonBox, &QDialogButtonBox::accepted, this, &QDialog::accept);
   QObject::connect(buttonBox, &QDialogButtonBox::rejected, this, &QDialog::reject);
 }
 
-void SeriesSelector::add(QListWidgetItem *item) {
+void SignalSelector::add(QListWidgetItem *item) {
   auto it = (ListItem *)item;
   addItemToList(selected_list, it->msg_id, it->sig, true);
   delete item;
 }
 
-void SeriesSelector::remove(QListWidgetItem *item) {
+void SignalSelector::remove(QListWidgetItem *item) {
   auto it = (ListItem *)item;
   if (it->msg_id == msgs_combo->currentData().value<MessageId>()) {
     addItemToList(available_list, it->msg_id, it->sig);
@@ -67,20 +77,20 @@ void SeriesSelector::remove(QListWidgetItem *item) {
   delete item;
 }
 
-void SeriesSelector::updateAvailableList(int index) {
+void SignalSelector::updateAvailableList(int index) {
   if (index == -1) return;
   available_list->clear();
   MessageId msg_id = msgs_combo->itemData(index).value<MessageId>();
   auto selected_items = seletedItems();
   for (auto s : dbc()->msg(msg_id)->getSignals()) {
-    bool is_selected = std::any_of(selected_items.begin(), selected_items.end(), [=, sig=s](auto it) { return it->msg_id == msg_id && it->sig == sig; });
+    bool is_selected = std::any_of(selected_items.begin(), selected_items.end(), [=, sig = s](auto it) { return it->msg_id == msg_id && it->sig == sig; });
     if (!is_selected) {
       addItemToList(available_list, msg_id, s);
     }
   }
 }
 
-void SeriesSelector::addItemToList(QListWidget *parent, const MessageId id, const cabana::Signal *sig, bool show_msg_name) {
+void SignalSelector::addItemToList(QListWidget *parent, const MessageId id, const cabana::Signal *sig, bool show_msg_name) {
   QString text = QString("<span style=\"color:%0;\">■ </span> %1").arg(getColor(sig).name(), sig->name);
   if (show_msg_name) text += QString(" <font color=\"gray\">%0 %1</font>").arg(msgName(id), id.toString());
 
@@ -91,8 +101,8 @@ void SeriesSelector::addItemToList(QListWidget *parent, const MessageId id, cons
   parent->setItemWidget(new_item, label);
 }
 
-QList<SeriesSelector::ListItem *> SeriesSelector::seletedItems() {
-  QList<SeriesSelector::ListItem *> ret;
+QList<SignalSelector::ListItem *> SignalSelector::seletedItems() {
+  QList<SignalSelector::ListItem *> ret;
   for (int i = 0; i < selected_list->count(); ++i) ret.push_back((ListItem *)selected_list->item(i));
   return ret;
 }
