+++ tools/cabana/mainwin.h
@@ -7,7 +7,7 @@
 #include <QSplitter>
 #include <QStatusBar>
 
-#include "tools/cabana/chartswidget.h"
+#include "tools/cabana/chart/chartswidget.h"
 #include "tools/cabana/dbc/dbcmanager.h"
 #include "tools/cabana/detailwidget.h"
 #include "tools/cabana/messageswidget.h"
